Fixed pixel updating
When I made each glyph a single table, changing the number or rows or columns would screw up the bitmap because the pixel
array size wasn't getting updated. Also, changing a pixel would not get saved in the pixel array. So I added an updatePixel
function which would copy out the pixel array from the table, update the width and/or height and then write the result to
the data_pixels string in the glyph. Now you can change the width and height, modify pixels, write the result and have it
correctly read back in the changes