Fixed bug where sometimes trailing commas are not output for GFXGlyph array
Sometimes it was miscalculating when the last line was being output so one or more commas was omitted
resulting in a syntax error. I didn't actually track down the source of the miscalculation. I just changed the
output to generate a comma at the start of each line, except for the first. Much easier to determine the
first line than the last